Major shooting rework
The way shots are handled has been reworked. FiredShot now takes a lot of the responsibility for calculating velocity. This helps put all of the velocity calculations in one place. In addition, damage has been changed. The gun's damage value is the base value. Attachments and bullets only provide multipliers.